Remove leftover naming utils. Cut release
